% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_custom_vcf_to_uniquorn_db.R
\name{add_custom_vcf_to_uniquorn_db}
\alias{add_custom_vcf_to_uniquorn_db}
\title{Adds a custom vcf file to the three existing cancer cell line panels
It is strongly recommended to filter the SNPs with a minor allel frequency of more than 0.01.}
\usage{
add_custom_vcf_to_uniquorn_db(vcf_file_path, ref_gen = "HG19", name_cl = "",
  safe_mode = T)
}
\description{
Adds a custom vcf file to the three existing cancer cell line panels
It is strongly recommended to filter the SNPs with a minor allel frequency of more than 0.01.
}

