% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_custom_vcfs_to_database.R
\name{add_custom_vcfs_to_database}
\alias{add_custom_vcfs_to_database}
\title{Adds a custom vcf files to the three existing cancer cell line panels}
\usage{
add_custom_vcfs_to_database( 
 vcf_input_files, 
ref_gen = "GRCH37", 
library = "", 
test_mode = FALSE,
n_threads = 1)
}
\arguments{
\item{vcf_input_files}{Input vcf file.s This may be one or many vcf files}

\item{ref_gen}{Reference genome version. All training sets are associated 
with a reference genome version. Default: GRCH37}

\item{library}{The name of the library to add the CCLs to. Standard is '_CUSTOM' 
will automatically be added as suffix.}

\item{test_mode}{Is this a test? Just for internal use}

\item{n_threads}{Specifies number of threads to be used}
}
\value{
Message if the adding has succeeded
}
\description{
Adds a custom vcf files to the three existing cancer cell line panels
}
\examples{
HT29_vcf_file = system.file("extdata/HT29.vcf.gz", package="Uniquorn");
add_custom_vcfs_to_database( 
vcf_input_files = HT29_vcf_file,
library = "",
ref_gen = "GRCH37",
test_mode = TRUE,
n _threads = 1)
}
