% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics_identification.R
\name{add_p_q_values_statistics}
\alias{add_p_q_values_statistics}
\title{add_p_q_values_statistics}
\usage{
add_p_q_values_statistics(g_query, match_t, p_value, ref_gen,
  minimum_matching_mutations, top_hits_per_library)
}
\arguments{
\item{g_query}{IRanges object that contains the query variants}

\item{match_t}{A table that contains the nubmber of matching variants}

\item{p_value}{Threshold for the significance p-value}

\item{ref_gen}{Reference genome version}

\item{minimum_matching_mutations}{Manual lower amount of matching
mutations require for a significant match between a query and a 
reference}

\item{top_hits_per_library}{limits significant similarities to the first 
n hits}
}
\value{
R table with a statistic
}
\description{
A hypergeometric distribution-assumption allows to calculate the 
p-values for a significant or non-significant overlap in this function
}
\details{
\code{add_p_q_values_statistics} Calculates the p-values
}
