% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_single_file.R
\name{add_single_file}
\alias{add_single_file}
\title{Build Fingerprint From VCF File}
\usage{
add_single_file(vcf_file_path, sim_list_stats, cl_id)
}
\arguments{
\item{vcf_file_path}{character string giving the path to the vcf file
on the operating system.}

\item{sim_list_stats}{table containing all cancer cell lines currently
contained in the database.}

\item{cl_id}{character string giving the cancer cell line identifier.}
}
\value{
Loci-based DNA-mutational fingerprint of the cancer cell line
 associated with the respective identifier.
}
\description{
Intern utility function. Wrapper function to filter VCF file using 
\code{parse_vcf_file} and build fingerprint with identifier.
}
