% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_unique_fingrprint_in_db.R
\name{create_mutation_fingerprint}
\alias{create_mutation_fingerprint}
\title{Create the unique fingerprint signature in the model data
export}
\usage{
create_mutation_fingerprint()
}
\description{
Create the unique fingerprint signature in the model data
export
}

