% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_signatures.R
\name{create_sim_list}
\alias{create_sim_list}
\title{Create similarity matrix from raw_data}
\usage{
create_sim_list(fingerprint_data, cl_data, type = "", panel = "")
}
\description{
Create similarity matrix from raw_data
}

