% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_similarity_matrix.R
\name{create_similarity_matrix}
\alias{create_similarity_matrix}
\title{Create similarity matrix from raw_data}
\usage{
create_similarity_matrix()
}
\description{
Create similarity matrix from raw_data
}

