% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_weights.R
\name{filter_for_weights}
\alias{filter_for_weights}
\title{filter_for_weights}
\usage{
filter_for_weights( 
    mutational_weight_inclusion_threshold,
    ref_gen,
    verbose,
    sim_list,
    sim_list_stats)
}
\arguments{
\item{mutational_weight_inclusion_threshold}{Lower bound for mutational weight to be 
included}

\item{ref_gen}{Reference genome version. All training sets are 
associated with a reference genome version. Default: GRCH37}

\item{verbose}{Print additional information}

\item{sim_list}{Contains the mutations}

\item{sim_list_stats}{Contains the overal mutation statistics}
}
\value{
Filtered reference sets
}
\description{
Filter the reference set
}
\details{
\code{filter_for_weights} parses vcf file and output 
basic information
}
