% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{identify_cancer_cell_line}
\alias{identify_cancer_cell_line}
\title{Start the identification of a cancer cell lines based on a vcf file}
\usage{
identify_cancer_cell_line(vcf_file_path)
}
\description{
Start the identification of a cancer cell lines based on a vcf file
}

