% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_VCF_file.R
\name{identify_vcf_file}
\alias{identify_vcf_file}
\title{Parses the vcf file and predicts the identity of the sample}
\usage{
identify_vcf_file(vcf_file, output_file = "", ref_gen = "HG19",
  unique_mode = F)
}
\description{
Parses the vcf file and predicts the identity of the sample
}

