% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_and_load_identification.R
\name{init_and_load_identification}
\alias{init_and_load_identification}
\title{init_and_load_identification}
\usage{
init_and_load_identification( 
verbose,
ref_gen,
vcf_file,
output_file,
n_threads)
}
\arguments{
\item{verbose}{Print additional information}

\item{ref_gen}{Reference genome version. All training sets are 
associated with a reference genome version. Default: GRCH37}

\item{vcf_file}{Path to vcf_file}

\item{output_file}{Path to output report file}

\item{n_threads}{Specifies number of threads to be used}
}
\value{
Three file path instances and the fingerprint
}
\description{
Initiate the analysis
Output basic information
}
\details{
\code{init_and_load_identification} parses vcf file and output 
basic information
}
