% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_data.R
\name{initiate_uniquorn_database}
\alias{initiate_uniquorn_database}
\title{Parses data into r list variable}
\usage{

  initiate_uniquorn_database(cosmic_genotype_file = "CosmicCLP_MutantExport.tsv",
  cellminer_genotype_file = "DNA__Exome_Seq_none.txt",
  ccle_genotype_file = "CCLE_hybrid_capture1650_hg19_allVariants_2012.05.07.maf",
  ucsc_db_snp_file = "snp142Common.txt", ref_gen = "hg19", weighted = T)
}
\description{
Parses data into r list variable
}

