% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_data.R
\name{initiate_younikorn_database}
\alias{initiate_younikorn_database}
\title{Parses data into the sql database
Requires the currently (Nov 2015) github located package "chapmandu2/CancerCellLines"}
\usage{
initiate_younikorn_database(parser_path)
}
\description{
Parses data into the sql database
Requires the currently (Nov 2015) github located package "chapmandu2/CancerCellLines"
}

