% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_younikorn_dependencies.R
\name{install_younikorn_dependencies}
\alias{install_younikorn_dependencies}
\title{Install packages required by Younikorn}
\usage{
install_younikorn_dependencies()
}
\description{
Install packages required by Younikorn
}

