% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_similarity_data_into_db.R
\name{load_similarity_data_into_db}
\alias{load_similarity_data_into_db}
\title{Loads similarity data into db}
\usage{
load_similarity_data_into_db(similarity_matrix_data, db_path = system.file("",
  package = "Younikorn"))
}
\description{
Loads similarity data into db
}

