% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_scripts.R
\name{parse_ccle_genotype_data}
\alias{parse_ccle_genotype_data}
\title{Parses ccle genotype data}
\usage{
parse_ccle_genotype_data(ccle_file, sim_list)
}
\arguments{
\item{ccle_file}{path to ccle file}

\item{sim_list}{Variable containing mutations and cell line}
}
\description{
Parses ccle genotype data
}

