% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse_CoSMIC_data.R
\name{parse_cosmic_clp_data}
\alias{parse_cosmic_clp_data}
\title{Parse CoSMIC CLP data
CosmicCLP_CompleteExport.tsv',}
\usage{
parse_cosmic_clp_data(parser_path, raw_data)
}
\description{
Parse CoSMIC CLP data
CosmicCLP_CompleteExport.tsv',
}

