% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_scripts.R
\name{parse_cosmic_genotype_data}
\alias{parse_cosmic_genotype_data}
\title{Parses cosmic genotype data}
\usage{
parse_cosmic_genotype_data(cosmic_file, sim_list)
}
\arguments{
\item{cosmic_file}{Path to cosmic clp file in hard disk}

\item{sim_list}{Variable containing mutations & cell line}
}
\description{
Parses cosmic genotype data
}

