% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_data.R
\name{parse_data_into_younikorn_db}
\alias{parse_data_into_younikorn_db}
\title{Parses data into the sql database
Requires the currently (Nov 2015) github located package "chapmandu2/CancerCellLines"}
\usage{
parse_data_into_younikorn_db(parser_path, db_path = system.file("database",
  package = "Younikorn"))
}
\description{
Parses data into the sql database
Requires the currently (Nov 2015) github located package "chapmandu2/CancerCellLines"
}

