% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_scripts.R
\name{parse_vcf_file}
\alias{parse_vcf_file}
\title{Filter Parsed VCF Files}
\usage{
parse_vcf_file(
    vcf_file,
    ref_gen,
    library_name
)
}
\arguments{
\item{vcf_file}{character string giving the path to the vcf file
on the operating system.}
}
\value{
Loci-based DNA-mutational fingerprint of the cancer cell line
 as found in the input VCF file.identify_vcf_files
}
\description{
Intern utility function. Filters the parsed VCF file for all informations
except for the start and length of variations/mutations.
}
