% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_custom_vcf_from_uniquorn.R
\name{remove_cl_from_uniquorn_db}
\alias{remove_cl_from_uniquorn_db}
\title{Removes a custom vcf file to the three existing cancer cell line panels}
\usage{
remove_cl_from_uniquorn_db(cl_name, ref_gen = "hg19", name_vcf = "")
}
\description{
Removes a custom vcf file to the three existing cancer cell line panels
}

