% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_database_information.R
\name{show_contained_cls}
\alias{show_contained_cls}
\title{Show all cancer cell line identifier present in the database for a selected reference Genome}
\usage{
show_contained_cls(ref_gen = "GRCH37", distinct_mode = T)
}
\description{
Show all cancer cell line identifier present in the database for a selected reference Genome
}

