% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_database_information.R
\name{show_contained_mutations}
\alias{show_contained_mutations}
\title{Show all mutations present in the database for a selected reference Genome}
\usage{
show_contained_mutations(ref_gen = "HG19")
}
\description{
Show all mutations present in the database for a selected reference Genome
}

