% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_database_information.R
\name{show_contained_mutations_for_cl}
\alias{show_contained_mutations_for_cl}
\title{Show all mutations present in the database for a selected cancer cell line and reference Genome}
\usage{
show_contained_mutations_for_cl(cl_name, ref_gen = "HG19")
}
\description{
Show all mutations present in the database for a selected cancer cell line and reference Genome
}

