% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser_scripts.R
\name{split_add_parallel}
\alias{split_add_parallel}
\title{split_add_parallel}
\usage{
split_add_parallel(para_index, vcf_matrix_row, vcf_handle, MARGIN, n_threads)
}
\arguments{
\item{para_index}{row of the vcf file}

\item{vcf_matrix_row}{A row of the parsed vcf file}

\item{vcf_handle}{Handle to the parsed VCF file}

\item{MARGIN}{Margin of the parse operation}

\item{n_threads}{Specifies number of threads to be used}
}
\value{
Transformed entry of vcf file, reduced to start and length
}
\description{
split_add_parallel
}
