% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_VCF_file.R
\name{identify_vcf_file}
\alias{identify_vcf_file}
\title{Parses the vcf file and predicts the identity of the sample}
\usage{
identify_vcf_file(vcf_file_path, output_path = "", panels = c("CELLMINER",
  "CCLE", "COSMIC"), ref_gen = "hg19")
}
\description{
Parses the vcf file and predicts the identity of the sample
}

