% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_VCF_file.R
\name{identify_vcf_file}
\alias{identify_vcf_file}
\title{Parses the vcf file and predicts the identity of the sample}
\usage{
identify_vcf_file(vcf_file, output_file = "", ref_gen = "GRCH37",
  similarity_threshold = 5, mutational_weight_inclusion_threshold = 1,
  only_first_candidate = FALSE, distinct_mode = TRUE, batch_mode = FALSE,
  write_xls = FALSE)
}
\arguments{
\item{vcf_file}{Input vcf file. Only one sample column allowed.}

\item{output_file}{Path of the output file. If blank, autogenerated as name of input file plus '_uniquorn_ident.tab' suffix.}

\item{ref_gen}{Reference genome version. All training sets are associated with a reference genome version. Default: GRCH37}

\item{similarity_threshold}{Cl identification that surpass this percentage of training mutation weight (importance) are predicted as present in the sample.}

\item{mutational_weight_inclusion_threshold}{Include only mutations with a weight of at least x. Range: 0.0 to 1.0. 1= unique to CL. ~0 = found in many CL samples.}

\item{only_first_candidate}{Only the CL identifier with highest score is predicted to be present in the sample}

\item{distinct_mode}{Show training data for the commonly or separately normalized training sets. Options: TRUE/ FALSE}

\item{batch_mode}{When many vcf files are to be analyzed in the same R session/ namespace with identical parameters, setting the parameter TRUE leads to a significant speed-up of the analysis.}

\item{write_xls}{Create identification results additionally as xls file for easier reading}
}
\value{
R table with a statistic of the identification result
}
\description{
Parses the vcf file and predicts the identity of the sample
}
\examples{
HT29_vcf_file = system.file("extdata/HT29.vcf.gz", package="Uniquorn"); identify_vcf_file( HT29_vcf_file )
}

