% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse_raw_data.R
\name{initiate_canonical_databases}
\alias{initiate_canonical_databases}
\title{Parses data into r list variable}
\usage{

  initiate_canonical_databases(cosmic_genotype_file = "CosmicCLP_MutantExport.tsv",
  ccle_genotype_file = "CCLE_hybrid_capture1650_hg19_allVariants_2012.05.07.maf",
  ref_gen = "HG19")
}
\description{
Parses data into r list variable
}

