% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_data.R
\name{initiate_younikorn_database}
\alias{initiate_younikorn_database}
\title{Parses data into r list variable}
\usage{
initiate_younikorn_database(parser_path)
}
\description{
Parses data into r list variable
}

