% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_similarity_data_into_db.R
\name{load_similarity_data_into_db}
\alias{load_similarity_data_into_db}
\title{Loads similarity data into db}
\usage{
load_similarity_data_into_db(sim_list, fingerprint_data, db_path)
}
\description{
Loads similarity data into db
}

