% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse_CCLE_hybrid_data.R
\name{parse_ccle_hybrid_data}
\alias{parse_ccle_hybrid_data}
\title{Parse the CCLE hybrid capture data}
\usage{
parse_ccle_hybrid_data(path_to_raw_data)
}
\description{
Parse the CCLE hybrid capture data
}

