% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_database_information.R
\name{show_contained_cls}
\alias{show_contained_cls}
\title{Display Contained Cancer Cell Lines}
\usage{
show_contained_cls(ref_gen, library = NULL)
}
\arguments{
\item{ref_gen}{a character vector specifying the reference genome version. All training sets are associated with a reference genome version. Default is \code{"GRCH37"}.}

\item{verbose}{Should DB informations be printed}
}
\value{
R table which contains the identifiers of all cancer cell line samples
 which match the specified parameters (reference genome and library).
}
\description{
This function displays the names, amount of mutations/variations and the overall weight of the mutations of all contained cancer cell line fingerprints 
for a chosen reference genome and optional library.
}
\examples{
##Show all contained cancer cell lines for reference GRCH37
show_contained_cls(ref_gen = "GRCH37", verbose = TRUE)

##Show just cancer cell lines contained in library CELLMINER
show_contained_cls(ref_gen = "GRCH37", library = "CELLMINER")
}
