% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_database_information.R
\name{show_contained_cls}
\alias{show_contained_cls}
\title{show_contained_cls}
\usage{
show_contained_cls( 
ref_gen)
}
\arguments{
\item{ref_gen}{Reference genome version. All training sets are associated with a reference genome version. Default: GRCH37}
}
\value{
R table which contains the identifier of all cancer cell line samples with the specific reference genome and the weight of all mutations
}
\description{
Show all cancer cell line identifier present in the database for a selected reference genome:
This function shows the names, amount of mutations/ variations, overall weight of the mutations of all contained training CLs 
for a chosen reference genome.
}
\examples{
contained_cls = show_contained_cls( 
ref_gen = "GRCH37")
}
