% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_database_information.R
\name{show_contained_mutations}
\alias{show_contained_mutations}
\title{Show all mutations present in the database for a selected reference Genome}
\usage{
show_contained_mutations( ref_gen = "GRCH37", distinct_mode = TRUE )
}
\arguments{
\item{ref_gen}{Reference genome version}

\item{distinct_mode}{Show mutations for either distinct or non-distinct normalization of mutational weights}
}
\value{
R Table which contains all mutations associated with a particular cancer cell line for a specified reference genome
}
\description{
This function shows all training-set mutations for a selected reference genome, i.e. the mutations that are being used
for identification of query cancer cell lines.
}

