% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_database_information.R
\name{show_contained_mutations}
\alias{show_contained_mutations}
\title{All Mutations For Reference Genome}
\usage{
show_contained_mutations(ref_gen)
}
\arguments{
\item{ref_gen}{a character vector specifying the reference genome version. All training sets are associated with a reference genome version. Default is \code{"GRCH37"}.}
}
\value{
R Table which contains all mutations associated with a specified reference genome.
}
\description{
This function shows all training-set mutations for a selected reference genome, i.e. the mutations that are being used
for identification of a query cancer cell line.
}
\examples{
show_contained_mutations(ref_gen = "GRCH37")
}
