% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_database_information.R
\name{show_contained_mutations_for_cl}
\alias{show_contained_mutations_for_cl}
\title{Mutations In Cancer Cell Line}
\usage{
show_contained_mutations_for_cl(name_cl, 
ref_gen)
}
\arguments{
\item{name_cl}{a character vector giving the identifier of the cancer cell line for which mutations will be shown.}

\item{ref_gen}{a character vector specifying the reference genome version. All training sets are associated with a reference genome version. Default is \code{"GRCH37"}.}
}
\value{
R table which contains all mutations associated with the defined cancer cell line and reference genome.
}
\description{
This function shows all mutations present in the database for a selected cancer cell line and reference genome.
}
\examples{
show_contained_mutations_for_cl(name_cl = "SK_OV_3_CELLMINER_mutations",
                                ref_gen = "GRCH37")
}
